Create docker sub-modules for more organization
